#0018287 Add new feature -> Collaboration ActionButtons are conditionally rendered according to current collaboration status and userType (freelance, employer). Each button triggers different action.

+++ client/fastwork-ui/src/components/chat/ChatRoom/partials/DetailsHeader/ActionButtons/freelancerActionButtonsMap/Enquiring.tsx
@@ -15,7 +15,9 @@ const Enquiring = () => {
 
     const handleOnClickDecline = () => {
         if(activeChatRoom){
-            deleteChatRoom(activeChatRoom?.id);
+            if(confirm('Are you sure to decline?')){
+                deleteChatRoom(activeChatRoom?.id);
+            }
         }
     }
